Moved util/Simd.h to simd/Simd.h and fixed license
