+++ openvdb/openvdb/CMakeLists.txt
@@ -601,6 +601,10 @@ set(OPENVDB_LIBRARY_THREAD_INCLUDE_FILES
   thread/Threading.h
 )
 
+set(OPENVDB_LIBRARY_SIMD_INCLUDE_FILES
+  simd/Simd.h
+)
+
 list(APPEND OPENVDB_LIBRARY_MATH_INCLUDE_FILES
   ${CMAKE_SOURCE_DIR}/ext/imath/openvdb/math/Half.h)
 list(APPEND OPENVDB_LIBRARY_SOURCE_FILES
@@ -811,6 +815,7 @@ install(FILES ${OPENVDB_LIBRARY_TOOLS_IMPL_INCLUDE_FILES} DESTINATION ${OPENVDB_
 install(FILES ${OPENVDB_LIBRARY_TREE_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/openvdb/tree)
 install(FILES ${OPENVDB_LIBRARY_UTIL_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/openvdb/util)
 install(FILES ${OPENVDB_LIBRARY_THREAD_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/openvdb/thread)
+install(FILES ${OPENVDB_LIBRARY_SIMD_INCLUDE_FILES} DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR}/openvdb/simd)
 
 if(USE_VCL)
   install(DIRECTORY ${CMAKE_SOURCE_DIR}/ext/vcl/openvdb DESTINATION ${OPENVDB_INSTALL_INCLUDEDIR})
